vaults move: ../data_sources/[dsId]/tables/csv
