+++ front/pages/api/v1/w/[wId]/data_sources/[dsId]/tables/csv.ts
@@ -1,79 +1,7 @@
-import type { NextApiRequest, NextApiResponse } from "next";
-
-import { withPublicAPIAuthentication } from "@app/lib/api/wrappers";
-import type { Authenticator } from "@app/lib/auth";
-import { DataSourceResource } from "@app/lib/resources/data_source_resource";
-import { apiError } from "@app/logger/withlogging";
-import { handleDataSourceTableCSVUpsert } from "@app/pages/api/w/[wId]/data_sources/[dsId]/tables/csv";
-
-export const config = {
-  api: {
-    bodyParser: {
-      sizeLimit: "50mb",
-    },
-  },
-};
+import handler from "@app/pages/api/v1/w/[wId]/vaults/[vId]/data_sources/[dsId]/tables/csv";
 
 /**
  * @ignoreswagger
- * System API key only endpoint. Undocumented.
+ * Legacy endpoint.
  */
-
-async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse,
-  auth: Authenticator
-): Promise<void> {
-  if (!auth.isSystemKey()) {
-    return apiError(req, res, {
-      status_code: 404,
-      api_error: {
-        type: "data_source_not_found",
-        message: "The data source you requested was not found.",
-      },
-    });
-  }
-
-  const { dsId } = req.query;
-  if (typeof dsId !== "string") {
-    return apiError(req, res, {
-      status_code: 400,
-      api_error: {
-        type: "invalid_request_error",
-        message: "Invalid path parameters.",
-      },
-    });
-  }
-
-  const dataSource = await DataSourceResource.fetchByNameOrId(
-    auth,
-    dsId,
-    // TODO(DATASOURCE_SID): Clean-up
-    { origin: "v1_data_sources_tables_csv" }
-  );
-  if (!dataSource) {
-    return apiError(req, res, {
-      status_code: 404,
-      api_error: {
-        type: "data_source_not_found",
-        message: "The data source you requested was not found.",
-      },
-    });
-  }
-
-  switch (req.method) {
-    case "POST":
-      return handleDataSourceTableCSVUpsert({ auth, req, res, dataSource });
-
-    default:
-      return apiError(req, res, {
-        status_code: 405,
-        api_error: {
-          type: "method_not_supported_error",
-          message: "The method passed is not supported, POST is expected.",
-        },
-      });
-  }
-}
-
-export default withPublicAPIAuthentication(handler);
+export default handler;
